Implemented conversion from Integer.

+++ src/powers_of_ten.rs
@@ -7,7 +7,7 @@
 // $Source$
 // $Revision$
 
-const POWERS_OF_10: [u128; 39] = [
+const POWERS_OF_10: [i128; 39] = [
     1,
     10,
     100,
@@ -50,11 +50,11 @@ const POWERS_OF_10: [u128; 39] = [
 ];
 
 #[inline]
-pub(crate) fn ten_pow(n: u8) -> u128 {
+pub(crate) fn ten_pow(n: u8) -> i128 {
     POWERS_OF_10[n as usize]
 }
 
 #[inline]
 pub(crate) fn checked_mul_pow_ten(val: i128, n: u8) -> Option<i128> {
-    val.checked_mul(ten_pow(n) as i128)
+    val.checked_mul(ten_pow(n))
 }
